Extract `FavoritePlayersPane` component out of `App`, passing `favoritePlayers` as a typed prop
