+++ players-forum/src/App.vue
@@ -1,15 +1,9 @@
 <script lang="ts">
 import { defineComponent } from 'vue';
+import type { Player } from '@/types';
+import FavoritePlayersPane from './components/FavoritePlayersPane.vue';
 
 const errorGettingPlayers = Math.random() > 1; // Change to simulate intermittent errors
-
-type Player = {
-  name: string;
-  team: string;
-  position: string;
-  number?: number;
-  captaincy: string;
-};
 
 const players: Player[] = errorGettingPlayers
   ? []
@@ -32,6 +26,7 @@ const players: Player[] = errorGettingPlayers
     ];
 
 export default defineComponent({
+  components: { FavoritePlayersPane },
   data() {
     return {
       favoriteTeam: 'Pittsburgh',
@@ -46,82 +41,10 @@ export default defineComponent({
       },
     };
   },
-  // Each computed property function begins with a log to easily see when they run
-  computed: {
-    // Something seems wrong about previous not having an inferred type
-    favoriteCaptains(previous: { favoriteCaptains: Player[] }): Player[] {
-      // This depends on favoritePlayers and runs every time a favorite is added
-      console.log('Recomputing favorite captains');
-      const newList = this.favoritePlayers.filter((player) => player.captaincy === 'C');
-
-      // It seems like previous is a proxy to an object containing all computeds/methods,
-      // so we really want to compare/return `previous.favoriteCaptains` and not `previous`.
-      // Is this inconsistent with the docs or am I reading them wrong?
-      //  https://vuejs.org/guide/essentials/computed.html#previous
-      console.log({ previousList: { ...previous.favoriteCaptains }, newList });
-
-      // Preserve the old reference if list length hasn't changed so dependents won't recompute
-      // Depending on the application, a more thorough deep equality check may be necessary
-      if (previous.favoriteCaptains && previous.favoriteCaptains.length === newList.length) {
-        console.log('Preserving previous favorite captains list');
-        return previous.favoriteCaptains;
-      }
-
-      console.log('Updated favorite captains list');
-      return newList;
-    },
-    totalFavoriteCaptains() {
-      // This depends on favoriteCaptains and runs every time a captain is added.
-      // It does not run if a non-captain is added, because favoriteCaptains will keep the same reference
-      console.log('Counting favorite captains');
-      return this.favoriteCaptains.length;
-    },
-    doubleTotalFavoriteCaptains() {
-      // This depends on totalFavoriteCaptains and runs every time a captain is added
-      // It does not run if a non-captain is added, because totalFavoriteCaptains remains unchanged
-      console.log('Doubling totalFavoriteCaptains');
-      return this.totalFavoriteCaptains * 2;
-    },
-    favoriteAlternates(): Player[] {
-      // This depends on favoritePlayers and runs every time a favorite is added
-      console.log('Recomputing favorite alternates');
-      return this.favoritePlayers.filter((player) => player.captaincy === 'A');
-    },
-    totalFavoriteAlternates() {
-      // This depends on favoriteAlternates and runs every time a favorite is added
-      // It runs even if a non-alternate is added because favoriteAlternates will have a new reference
-      console.log('Counting favorite alternates');
-      return this.favoriteAlternates.length;
-    },
-    doubleTotalFavoriteAlternates() {
-      // This depends on totalFavoriteAlternates and runs every time an alternate is added
-      // It does not run if a non-alternate is added, because totalFavoriteAlternates remains unchanged
-      console.log('Doubling totalFavoriteAlternates');
-      return this.totalFavoriteAlternates * 2;
-    },
-    // Calculating percents depends on the entire favorites array, so using those computed properties makes all others run again
-    // totalFavorites() {
-    //   // This depends on favoritePlayers and runs every time a favorite is added
-    //   console.log('Counting favorites');
-    //   return this.favoritePlayers.length;
-    // },
-    // favoritesPercentCaptains() {
-    //   console.log('Calculating % captains in favorites');
-    //   if (this.totalFavorites === 0) return 0; // Avoid divide by zero
-    //   return Math.round((this.totalFavoriteCaptains / this.totalFavorites) * 100);
-    // },
-    // favoritesPercentAlternates() {
-    //   console.log('Calculating % alternates in favorites');
-    //   if (this.totalFavorites === 0) return 0; // Avoid divide by zero
-    //   return Math.round((this.totalFavoriteAlternates / this.totalFavorites) * 100);
-    // },
-  },
   methods: {
     addPlayer(event: Event) {
       // A real app might validate the new player before adding it
-
       this.players.push(this.newPlayer);
-
       // Clear the form by setting the bound variable to an empty object
       this.newPlayer = {
         name: '',
@@ -130,7 +53,6 @@ export default defineComponent({
         number: undefined,
         captaincy: '',
       };
-
       // Move focus to the name field, found within the submitted form
       const nameInput: HTMLInputElement | null = (event.target as HTMLElement).querySelector(
         'input[name="name"]',
@@ -139,7 +61,6 @@ export default defineComponent({
     },
     addFavoritePlayer(player: Player) {
       // A real app might make sure the player isn't already a favorite, or store this data in a different format
-
       this.favoritePlayers.push(player);
     },
   },
@@ -209,23 +130,6 @@ export default defineComponent({
       </label>
       <button>Add player</button>
     </form>
-    <div v-if="favoritePlayers.length > 0" class="pane favorite-players-pane">
-      <h2>Favorite players</h2>
-      <dl>
-        <dt>Captains:</dt>
-        <dd>{{ totalFavoriteCaptains }} (double: {{ doubleTotalFavoriteCaptains }})</dd>
-        <!-- <dd>{{totalFavoriteCaptains}} ({{favoritesPercentCaptains}}%)</dd> -->
-
-        <dt>Alternates:</dt>
-        <dd>{{ totalFavoriteAlternates }} (double: {{ doubleTotalFavoriteAlternates }})</dd>
-        <!-- <dd>{{totalFavoriteAlternates}} ({{favoritesPercentAlternates}}%)</dd> -->
-      </dl>
-      <ul>
-        <li class="player player--favorite" v-for="player in favoritePlayers" :key="player.name">
-          <h3>{{ player.name }}</h3>
-          <aside>({{ player.team }})</aside>
-        </li>
-      </ul>
-    </div>
+    <FavoritePlayersPane :favorite-players="favoritePlayers" />
   </div>
 </template>
