+++ players-forum/src/types.ts
@@ -1,0 +1,7 @@
+export type Player = {
+  name: string;
+  team: string;
+  position: string;
+  number?: number;
+  captaincy: string;
+};
